Move full_with_rules_tests, rewrite w/ Pytest
We don't need the isolated file system as we're writing to the db.
